requirements.txt
flask
pdfminer.six
python-docx
beautifulsoup4
requests
fake-useragent
nltk
spacy
feedparser
